dont suport int8 onnx

+++ src/kabosu_plus/sbv2/nlp/korean/bert_feature.py
@@ -70,7 +70,7 @@ def extract_bert_feature_onnx(
     # text から BERT 特徴量を抽出
     io_binding.bind_output(output_name, device_type)
     session.run_with_iobinding(io_binding, run_options=run_options)
-    res = io_binding.get_outputs()[0].numpy().astype(np.float32)
+    res = io_binding.get_outputs()[0].numpy()
 
     style_res_mean = None
     if assist_text:
@@ -91,7 +91,7 @@ def extract_bert_feature_onnx(
         # assist_text から BERT 特徴量を抽出
         io_binding.bind_output(output_name, device_type)
         session.run_with_iobinding(io_binding, run_options=run_options)
-        style_res = io_binding.get_outputs()[0].numpy().astype(np.float32)
+        style_res = io_binding.get_outputs()[0].numpy()
         style_res_mean = np.mean(style_res, axis=0)
 
     assert len(word2ph) == len(text) + 2, text
